% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cal_rmse}
\alias{cal_rmse}
\title{Root Mean Squared Error (RMSE) of One Prediction
  Because we don't count the NA value in the following calculation,
  the length of the vec_pred must minus the NA prediction.
  Usually, there is no NA in prediction.}
\usage{
cal_rmse(vec_pred, vec_valid)
}
\description{
Root Mean Squared Error (RMSE) of One Prediction
  Because we don't count the NA value in the following calculation,
  the length of the vec_pred must minus the NA prediction.
  Usually, there is no NA in prediction.
}
