% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greybox.R
\name{est_mod_ctsm}
\alias{est_mod_ctsm}
\title{Estimate the model using data}
\usage{
est_mod_ctsm(mod, ti_data)
}
\arguments{
\item{mod}{CTMS Model}

\item{ti_data}{Tibble of data}
}
\description{
Estimate the model using data
}
