% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixTSA.R
\docType{package}
\name{MatrixTSA-package}
\alias{MatrixTSA}
\alias{MatrixTSA-package}
\title{MatrixTSA: Multivariate Time Series Analysis in Matrix}
\description{
Multivariate Time Series Analysis in Matrix
    Regression: Local Regression
    Test
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/edxu96/MatrixTSA}
}

}
\keyword{internal}
