% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_resi}
\alias{plot_resi}
\title{Model validation (check i.i.d. and distribution of residuals)}
\usage{
plot_resi(vec_resi, vec_in)
}
\arguments{
\item{vec_resi}{vector of residuals}

\item{vec_in}{vector of input}
}
\description{
Model validation (check i.i.d. and distribution of residuals)
}
