% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{plot_splines}
\alias{plot_splines}
\title{Plot the Fitted Splines Regression Model}
\usage{
plot_splines(li)
}
\arguments{
\item{li}{Fittted spline regression model}
}
\description{
Plot the Fitted Splines Regression Model
}
