% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splines.R
\name{reg_splines}
\alias{reg_splines}
\title{Fit a Splines Regression Model}
\usage{
reg_splines(li, df = 2, degree = 1, whe_intercept = TRUE)
}
\arguments{
\item{li}{List with x and y.}
}
\value{
Linear regression model
}
\description{
Fit a Splines Regression Model
}
