% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greybox.R
\name{set_mod_ctsm}
\alias{set_mod_ctsm}
\title{Set a new CTSM model}
\usage{
set_mod_ctsm(c_expr_sys, expr_obs, expr_error, c_input, datf_para)
}
\arguments{
\item{mat_bound}{Matrix of initial values and bounds for optimization}

\item{eval_input}{Evaluation to add input to the model}
}
\description{
Set a new CTSM model
}
