% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidynamics.R
\docType{package}
\name{tidynamics-package}
\alias{tidynamics}
\alias{tidynamics-package}
\title{tidynamics: Tidy Multivariate (Non)Linear Dynamic Systems}
\description{
Multivariate Regression of System Dynamics Tidily
    Regression: Local Regression
    Test
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/edxu96/tidynamics}
}

}
\keyword{internal}
