% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{trans_ctsm_ss}
\alias{trans_ctsm_ss}
\title{Transform the CTSM model to state space model}
\usage{
trans_ctsm_ss(mod_ctsm)
}
\arguments{
\item{mod_ctsm}{CTMS model to be tranformed}
}
\description{
Transform the CTSM model to state space model
}
