% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{trans_mat_ss}
\alias{trans_mat_ss}
\title{Transform the matrix for continuous state space model to matrix for
discerete state space model.}
\usage{
trans_mat_ss(li_mat_ss, ti_data)
}
\arguments{
\item{li_mat_ss}{List of matrix for continuous state space model}

\item{ti_data}{Tibble of data for estimation}
}
\description{
Transform the matrix for continuous state space model to matrix for
discerete state space model.
}
