% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{weight_tri}
\alias{weight_tri}
\title{Triangular Kernel Function
  x_i: the point for which to center
  x: the data points
  h: the bandwidth
  plot(x, tri(x_i = 0, x = x, h = 0.4))}
\usage{
weight_tri(x_i, x, h)
}
\description{
Triangular Kernel Function
  x_i: the point for which to center
  x: the data points
  h: the bandwidth
  plot(x, tri(x_i = 0, x = x, h = 0.4))
}
