% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greybox.R
\name{set_mod_ctsm}
\alias{set_mod_ctsm}
\title{Set a new CTSM model}
\usage{
set_mod_ctsm(c_expr_sys, expr_obs, expr_var, c_input, ti_est)
}
\arguments{
\item{c_expr_sys}{Vector of system expressions}

\item{expr_obs}{Observation expressions}

\item{expr_var}{Variance Expressions}

\item{c_input}{Name of Input Variables}

\item{datf_para}{Dataframe of estimated parameters}
}
\description{
Set a new CTSM model
}
